backend/requirements.txt
flask
unsloth
langchain-groq
python-dotenv
torch
transformers
xformers